requirements.txt
imgurpython==1.1.7
Pillow==10.1.0
